introduced ExcelPartitionReaderFromIterator which closes the workbook after read
